% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getlubriInt.R
\docType{methods}
\name{getlubriInt}
\alias{getlubriInt}
\alias{getlubriInt,RepoTimeInt-method}
\title{Return the slot lubriInt}
\usage{
getlubriInt(object)

\S4method{getlubriInt}{RepoTimeInt}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{RepoTimeInt}.}
}
\value{
A named list of objects of class Interval from package lubridate.
}
\description{
\code{getlubriInt} returns the slot \code{lubriInt} of an object
 of class \linkS4class{RepoTimeInt}.
}
\examples{
getlubriInt(newRepoTime('TT12015'))

}

