% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRepo.R
\docType{methods}
\name{getRepo}
\alias{getRepo}
\alias{getRepo,RepoTimeInt-method}
\title{Devuelve el slot Repo de un objeto de clase RepoTimeInt.}
\usage{
getRepo(object)

\S4method{getRepo}{RepoTimeInt}(object)
}
\arguments{
\item{object}{Objeto de clase \code{\linkS4class{RepoTimeInt}} del que se desea 
extraer el slot Repo.}
}
\value{
Un vector de clase \code{\link{character}} de longitud 1.
}
\description{
\code{getRepo} devuelve el slot \code{Repo} del input 
\code{\linkS4class{RepoTimeInt}}.
}
\details{
Esta función devuelve únicamente los datos del slot \code{Repo} del objeto 
\code{RepoTimeInt} de entrada como un vector.
}
\examples{
getRepo(new(Class = 'RepoTimeInt', 'TT12015'))

}

