% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Intersect.RepoTimeInt.R
\docType{methods}
\name{Intersect}
\alias{Intersect}
\alias{Intersect,RepoTimeInt-method}
\title{Method for S4 class \linkS4class{RepoTimeInt}}
\usage{
Intersect(x, y)

\S4method{Intersect}{RepoTimeInt}(x, y)
}
\arguments{
\item{x}{Object of class \linkS4class{RepoTimeInt}.}

\item{y}{Object of class \linkS4class{RepoTimeInt}.}
}
\value{
Object of class \linkS4class{RepoTimeInt} resulting from intersecting
the lubriInt slot(s) of the argument(s).
}
\description{
\code{Intersect} provides the intersection between two objects 
of class \linkS4class{RepoTimeInt} and length 1 or of all elements of an 
object of class \linkS4class{RepoTimeInt} and length greater than 1.
}
\examples{
RepoPeriod1 <- newRepoTime('MM012015')
RepoPeriod2 <- newRepoTime('MM022015')
Intersect(RepoPeriod1,RepoPeriod2)

}

