% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Length.R
\docType{methods}
\name{Length}
\alias{Length}
\alias{Length,RepoTimeInt-method}
\title{Length of an object of class \linkS4class{RepoTimeInt}}
\usage{
Length(x)

\S4method{Length}{RepoTimeInt}(x)
}
\arguments{
\item{x}{Object of class \linkS4class{RepoTimeInt}}
}
\value{
A non-negative integer.
}
\description{
\code{Length} gives the number of time intervals contained in an
object of class \linkS4class{RepoTimeInt}.
}
\examples{
RepoPeriod1 <- newRepoTime('MM012015')
Length(RepoPeriod1)

RepoPeriod2 <- newRepoTime(c('MM012015', 'MM022015'))
Length(RepoPeriod2)
}

