% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lubriToRepoTime.R
\name{lubriToRepoTime}
\alias{lubriToRepoTime}
\title{Transformation of \linkS4class{Interval} class objects into time 
intervals under the repo notation}
\usage{
lubriToRepoTime(lubriInterval, Rot = FALSE)
}
\arguments{
\item{lubriInterval}{Object of class \linkS4class{Interval} from 
\link[lubridate]{lubridate-package}.}

\item{Rot}{Logical vector of length 1 indicating whether periods with 
rotated samples are to be included in the sequence or not (default).}
}
\description{
\code{lubriToRepoTime} transforms time intervals expressed as 
objects of class \linkS4class{Interval} from the 
\link[lubridate]{lubridate-package} into the so-called repo notation.
}
\details{
The repo notation of time intervals is explained in the details 
section of \code{\link{RepoTimeInt-class}}.
}
\examples{
library(lubridate)
lubriToRepoTime(interval('2015-04-01', '2015-04-30', tz = 'Europe/Madrid'))
lubriToRepoTime(interval('2015-01-01', '2015-03-31', tz = 'Europe/Madrid'))

}
\seealso{
\code{\link{RepoTimeTolubri}}
}

