% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newRepoTime.R
\name{newRepoTime}
\alias{newRepoTime}
\title{Constructor de objetos de clase RepoTimeInt}
\usage{
newRepoTime(Time)
}
\arguments{
\item{Time}{un vector \code{character} con el período temporal siguiendo la 
notación del repositorio de microdatos.}
}
\value{
Objeto de clase \linkS4class{RepoTimeInt}.
}
\description{
Este constructor devuelve un objeto de clase 
\code{\linkS4class{RepoTimeInt}} a partir de un vector de períodos temporales
según la notación del repositorio de microdatos.
}
\examples{
x <- newRepoTime(c('MM042015', 'MM052015')) 
x
str(x) 

}

