% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orderRepoTime.R
\name{orderRepoTime}
\alias{orderRepoTime}
\title{Order of RepoTime Period}
\usage{
orderRepoTime(RepoTime)
}
\arguments{
\item{RepoTime}{Character vector with time periods in repo notation.}
}
\value{
An integer vector with the same length as the input parameter RepoTime with the order of 
the time periods in the input parameter.
}
\description{
\code{orderRepoTime} returns a character vector with the order of the time periods 
in the input parameter.
}
\examples{
orderRepoTime(paste0('MM', c('04','06', '01'), '2016'))
orderRepoTime(paste0('QQ', rep(c('1', '2'), 3),c('04','06', '01'), '2016'))
orderRepoTime(paste0('AA', c('2016', '2001', '2014', '2012')))

}
\seealso{
\code{\link{RepoTimeInt-class}}, \code{\link{getlubriInt}}
}
