% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plus.RepoTimeInt.R
\docType{methods}
\name{+,RepoTimeInt,Period-method}
\alias{+,RepoTimeInt,Period-method}
\title{Método '+' para la clase RepoTimeInt.}
\usage{
\S4method{+}{RepoTimeInt,Period}(e1, e2)
}
\arguments{
\item{e1}{Objeto de clase \code{\linkS4class{RepoTimeInt}}.}

\item{e2}{Objeto de clase \code{\linkS4class{lubridate::Period}}.}
}
\value{
Objeto de clase \code{\linkS4class{RepoTimeInt}} el resultado de sumar
al período inicial el plazo temporal especificado.
}
\description{
\code{+} suma a un objeto de clase \code{\linkS4class{RepoTimeInt}} un objeto
de clase \code{\linkS4class{lubridate::Period}} del paquete \code{lubridate}
con el plazo temporal de duración especificada como input.
}
\details{
Este método sobrecarga el operador \link{+} y construye un nuevo objeto
\code{RepoTimeInt}.
}
\examples{
RepoPeriod <- newRepoTime('MM022015')
RepoPeriod + months(1)
RepoPeriod + years(1)
RepoPeriod + weeks(1)

}

