% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DocPackage.R
\docType{package}
\name{RepoTime-package}
\alias{RepoTime-package}
\title{RepoTime: a package to manage time intervals usual in official statistics}
\description{
The package RepoTime provides an S4 class for time intervals expressed in the 
so-called repo notation for this package (see \code{\link{RepoTimeInt-class}}
 for details). This repo notation is intended to condense the time reference
 to data sets used to produce official statistics referred to by data 
 collection time period.
 
The package provides also usual methods for operating on these objects.
}

