% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minus.RepoTimeInt.R
\name{-,RepoTimeInt,Period-method}
\alias{-,RepoTimeInt,Period-method}
\title{Subtract of a \linkS4class{Period} class object to a 
\linkS4class{RepoTimeInt} class object}
\usage{
\S4method{-}{RepoTimeInt,Period}(e1, e2)
}
\arguments{
\item{e1}{Object of class \linkS4class{RepoTimeInt}.}

\item{e2}{Object of class \linkS4class{Period}.}
}
\value{
Object of class \linkS4class{RepoTimeInt} resulting from subtracting the
time period \code{e2} to the initial time interval \code{e1}.
}
\description{
\code{-} subtracts an object of class \linkS4class{Period} 
of \link[lubridate]{lubridate-package} to an object of class 
\linkS4class{RepoTimeInt}.
}
\details{
This method overloads the operator \link{-} and builds an object of
class \linkS4class{RepoTimeInt}.
}
\examples{
RepoPeriod <- newRepoTime('MM022015')
RepoPeriod - months(1)
RepoPeriod - lubridate::years(1)
RepoPeriod - lubridate::weeks(2)

}
