% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RepoTimeTolubri.R
\name{RepoTimeTolubri}
\alias{RepoTimeTolubri}
\title{Transformación de un intervalo temporal en notación del repositorio a 
del paquete lubridate
 
La notación de intervalos temporales del repositorio consiste en denotarlos
por la cadena de caracteres PPp...p, donde
\itemize{
 \item PP es de la forma 
         QQ,QR 	Quincenal
         MM,MR 	Mensual
         BB,BR 	Bimensual
         TT,TR 	Trimestral
         EE,ER 	Semestral
         AA,AR 	Anual
\item p...p es de la forma
        p….p 	Valores 	Periodicidad
        qmmaaaa  q=1,2;      mm=01,…,12; aaaa=año  Quincenal
        mmaaaa   mm=01,…,12; aaaa=año              Mensual
        baaaa    b=1,…,6;    aaaa=año 	          Bimensual
        taaaa    t=1,…,4;    aaaa=año 	          Trimestral
        saaaa 	s=1,2;      aaaa=año 	          Semestral
        aaaa     aaaa=año 	                      Anual
}}
\usage{
RepoTimeTolubri(RepoTime)
}
\description{
Transformación de un intervalo temporal en notación del repositorio a 
del paquete lubridate
 
La notación de intervalos temporales del repositorio consiste en denotarlos
por la cadena de caracteres PPp...p, donde
\itemize{
 \item PP es de la forma 
         QQ,QR 	Quincenal
         MM,MR 	Mensual
         BB,BR 	Bimensual
         TT,TR 	Trimestral
         EE,ER 	Semestral
         AA,AR 	Anual
\item p...p es de la forma
        p….p 	Valores 	Periodicidad
        qmmaaaa  q=1,2;      mm=01,…,12; aaaa=año  Quincenal
        mmaaaa   mm=01,…,12; aaaa=año              Mensual
        baaaa    b=1,…,6;    aaaa=año 	          Bimensual
        taaaa    t=1,…,4;    aaaa=año 	          Trimestral
        saaaa 	s=1,2;      aaaa=año 	          Semestral
        aaaa     aaaa=año 	                      Anual
}
}
\examples{
RepoTimeTolubri('MM022014')
RepoTimeTolubri('MM22014')

}

