% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seq.RepoTimeInt.R
\docType{methods}
\name{Seq}
\alias{Seq}
\alias{Seq,RepoTimeInt-method}
\title{Method for the S4 class \linkS4class{RepoTimeInt}}
\usage{
Seq(x, y, Rot = TRUE, RotPer = "12")

\S4method{Seq}{RepoTimeInt}(x, y, Rot = TRUE, RotPer = "12")
}
\arguments{
\item{x}{Object of class \linkS4class{RepoTimeInt} with the initial time 
interval.}

\item{y}{Object of class \linkS4class{RepoTimeInt} with the final time 
interval.}
}
\value{
A list of objects of class \linkS4class{RepoTimeInt} ranging from
the initial time interval until the final time interval.
}
\description{
\code{Seq} generates a list of objects of class 
\linkS4class{RepoTimeInt} ranging from the initial and final time interval 
arguments.
}
\examples{
RepoPeriod1 <- newRepoTime('MM012015')
RepoPeriod2 <- newRepoTime('MM022016')
Seq(RepoPeriod1, RepoPeriod2)
 
}

