% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.RepoTimeInt.R
\docType{methods}
\name{show,RepoTimeInt-method}
\alias{show,RepoTimeInt-method}
\title{Muestra un objeto de clase RepoTimeInt.}
\usage{
\S4method{show}{RepoTimeInt}(object)
}
\arguments{
\item{object}{Objeto de clase \code{\linkS4class{RepoTimeInt}} que se desea 
mostrar.}
}
\value{
Objeto de clase \code{\link{NULL}}.
}
\description{
\code{show.RepoTimeInt} muestra el slot \code{Repo} del input 
\code{\linkS4class{RepoTimeInt}}.
}
\details{
Esta función muestra únicamente los datos del slot \code{Repo} del objeto 
\code{RepoTimeInt} de entrada como un vector.

Se trata en realidad del método \code{\linkS4class{show}} adaptado para la 
clase \code{\linkS4class{RepoTimeInt}}.
}
\examples{
show(new(Class = 'RepoTimeInt', 'TT12015'))

}

