% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/subset.RepoTimeInt.R
\docType{methods}
\name{[,RepoTimeInt,ANY,ANY-method}
\alias{[,RepoTimeInt,ANY,ANY-method}
\title{Extract parts of an object of class \linkS4class{RepoTimeInt}}
\usage{
\S4method{[}{RepoTimeInt,ANY,ANY}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{Object of class \linkS4class{RepoTimeInt} from which to extract
parts.}

\item{i,j,...}{indices corresponding to the elements to extract.}

\item{drop}{Included for coherence.}
}
\value{
Object of class \linkS4class{RepoTimeInt}.
}
\description{
\code{[} extracts parts of an object of class
\linkS4class{RepoTimeInt}.
}
\examples{
Ejemplo <- newRepoTime(paste0('MM', c('03', '04'), '2014'))
Ejemplo[1]
}

