% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Union.RepoTimeInt.R
\docType{methods}
\name{Union}
\alias{Union}
\alias{Union,RepoTimeInt-method}
\title{Método 'Union' para la clase RepoTimeInt.}
\usage{
Union(x, y)

\S4method{Union}{RepoTimeInt}(x, y)
}
\arguments{
\item{x}{Objeto de clase \code{\linkS4class{RepoTimeInt}}.}

\item{y}{Objeto de clase \code{\linkS4class{RepoTimeInt}}.}
}
\value{
Objeto de clase \code{\linkS4class{RepoTimeInt}} el resultado de sumar
los periodos de tiempo.
}
\description{
\code{Union} suma dos objetos de clase \code{\linkS4class{RepoTimeInt}} y longitud 1
o los elementos de un objeto de clase \code{\linkS4class{RepoTimeInt}} y longitud
mayor o igual a 1.
}
\examples{
RepoPeriod1 <- newRepoTime('TT12015')
RepoPeriod2 <- newRepoTime('TT22015')
Union(RepoPeriod1,RepoPeriod2)

}

