% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getlubriInt.R
\docType{methods}
\name{getlubriInt}
\alias{getlubriInt}
\alias{getlubriInt,RepoTimeInt-method}
\title{Devuelve el slot lubriInt de un objeto de clase RepoTimeInt.}
\usage{
getlubriInt(object)

\S4method{getlubriInt}{RepoTimeInt}(object)
}
\arguments{
\item{object}{Objeto de clase \code{\linkS4class{RepoTimeInt}} del que se desea 
extraer el slot Repo.}
}
\value{
Una lista de objetos de clase \code{\link{lubridate::Interval}}.
}
\description{
\code{getlubriInt} devuelve el slot \code{lubriInt} del input 
\code{\linkS4class{RepoTimeInt}}.
}
\details{
Esta función devuelve únicamente los datos del slot \code{lubriInt} del objeto 
\code{RepoTimeInt} de entrada como un vector.
}
\examples{
getlubriInt(new(Class = 'RepoTimeInt', 'TT12015'))

}

