% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRepo.R
\name{getRepo}
\alias{getRepo}
\alias{getRepo,RepoTimeInt-method}
\title{Return the slot Repo}
\usage{
getRepo(object)

\S4method{getRepo}{RepoTimeInt}(object)
}
\arguments{
\item{object}{Object of class \code{\linkS4class{RepoTimeInt}}.}
}
\value{
A character vector with the time intervals in the repo notation.
}
\description{
\code{getRepo} returns the slot \code{Repo} of an object of 
class \linkS4class{RepoTimeInt}.
}
\examples{
getRepo(newRepoTime('TT12015'))


}
\seealso{
\code{\link{RepoTimeInt-class}}, \code{\link{getlubriInt}}
}
