% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Intersect.RepoTimeInt.R
\docType{methods}
\name{Intersect}
\alias{Intersect}
\alias{Intersect,RepoTimeInt-method}
\title{Método 'Intersect' para la clase RepoTimeInt.}
\usage{
Intersect(x, y)

\S4method{Intersect}{RepoTimeInt}(x, y)
}
\arguments{
\item{x}{Objeto de clase \code{\linkS4class{RepoTimeInt}}.}

\item{y}{Objeto de clase \code{\linkS4class{RepoTimeInt}}.}
}
\value{
Objeto de clase \code{\linkS4class{RepoTimeInt}} el resultado de hacer la
intersección de los periodos de tiempo.
}
\description{
\code{Intersect} da la intersección de dos objetos de clase \code{\linkS4class{RepoTimeInt}}
y longitud 1 o de los elementos de un objeto de clase \code{\linkS4class{RepoTimeInt}}
y longitud mayor o igual a 1.
}
\examples{
RepoPeriod1 <- newRepoTime('MM012015')
RepoPeriod2 <- newRepoTime('MM022015')
Intersect(RepoPeriod1,RepoPeriod2)

}

