% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Intersect.RepoTimeInt.R
\name{Intersect}
\alias{Intersect}
\alias{Intersect,RepoTimeInt-method}
\title{Intersect of two objects of class \linkS4class{RepoTimeInt}}
\usage{
Intersect(x, y)

\S4method{Intersect}{RepoTimeInt}(x, y)
}
\arguments{
\item{x}{Object of class \linkS4class{RepoTimeInt}.}

\item{y}{Object of class \linkS4class{RepoTimeInt}.}
}
\value{
Object of class \linkS4class{RepoTimeInt} resulting from intersecting
the slot(s) \code{lubriInt} of the argument(s).
}
\description{
\code{Intersect} provides the intersection between two objects 
of class \linkS4class{RepoTimeInt} and length 1 or of all elements of an 
object of class \linkS4class{RepoTimeInt} and length greater than 1.
}
\details{
If argument \code{y} is missing, \code{Intersect} is applied 
iteratively on every successive component of the list of 
\code{\link[lubridate]{Interval-class}} objects. If argument \code{y} is 
not missing, the length of both arguments must be 1.
}
\examples{
RepoPeriod1 <- newRepoTime('MM012015')
RepoPeriod2 <- newRepoTime('MM022015')
Intersect(RepoPeriod1, RepoPeriod2)

RepoPeriod1 <- newRepoTime('MM012015')
RepoPeriod2 <- newRepoTime('AA2015')
Intersect(RepoPeriod1, RepoPeriod2)

}
