% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plus.RepoTimeInt.R
\docType{methods}
\name{+,RepoTimeInt,Period-method}
\alias{+,RepoTimeInt,Period-method}
\title{Method for the S4 class \linkS4class{RepoTimeInt}}
\usage{
\S4method{+}{RepoTimeInt,Period}(e1, e2)
}
\arguments{
\item{e1}{Object of class \linkS4class{RepoTimeInt}.}

\item{e2}{Object of class \linkS4class{Period}.}
}
\value{
Object of class \linkS4class{RepoTimeInt} resulting from summing the
time period \code{y} to the initial time interval \code{x}.
}
\description{
\code{+} sums an object of class \linkS4class{Period} of package
lubridate to an object of class \linkS4class{RepoTimeInt}.
}
\details{
This method overloads the operator \link{+} and builds an object of
class \linkS4class{RepoTimeInt}.
}
\examples{
library(lubridate)
RepoPeriod <- newRepoTime('MM022015')
RepoPeriod + months(1)
RepoPeriod + years(1)
RepoPeriod + weeks(1)

}

