% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RepoTimeTolubri.R
\name{RepoTimeTolubri}
\alias{RepoTimeTolubri}
\title{Transformation of time intervals in repo notation into a list of
\link[lubridate]{Interval-class} objects of the 
\link[lubridate]{lubridate-package}}
\usage{
RepoTimeTolubri(RepoTime, TimeZone = "Europe/Madrid")
}
\arguments{
\item{RepoTime}{Character vector with the time interval in repo notation.}

\item{TimeZone}{Time zone component of a date-time as expected in function
\code{\link[lubridate]{tz}} of the \link[lubridate]{lubridate-package} 
(default value = 'Europe/Madrid').}
}
\description{
\code{RepoTimeTolubri} transforms time intervals expressed in 
the so-called repo notation (see details) into a list of 
\link[lubridate]{Interval-class} objects of the 
\link[lubridate]{lubridate-package}. When the time interval is not recognized
as a valid string, it returns the corresponding message.
}
\details{
The repo notation of time intervals is explained in the details 
section of \code{\link{RepoTimeInt-class}}.
}
\examples{
RepoTimeTolubri('MM022014')
RepoTimeTolubri('MR122014')
RepoTimeTolubri('TT20141')

}

