% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seq.RepoTimeInt.R
\docType{methods}
\name{Seq}
\alias{Seq}
\alias{Seq,RepoTimeInt-method}
\title{Método 'Seq' para la clase RepoTimeInt.}
\usage{
Seq(x, y)

\S4method{Seq}{RepoTimeInt}(x, y)
}
\arguments{
\item{x}{Objeto de clase \code{\linkS4class{RepoTimeInt}}.}

\item{y}{Objeto de clase \code{\linkS4class{RepoTimeInt}}.}
}
\value{
Objeto de clase \code{\linkS4class{RepoTimeInt}} el resultado de sumar
los periodos de tiempo.
}
\description{
\code{Seq} genera una colección de objetos de clase \code{\linkS4class{RepoTimeInt}}
}
\examples{
RepoPeriod1 <- newRepoTime('MM012015')
RepoPeriod2 <- newRepoTime('MM022016')
Seq(RepoPeriod1,RepoPeriod2)
 
}

