% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seq.RepoTimeInt.R
\docType{methods}
\name{Seq}
\alias{Seq}
\alias{Seq,RepoTimeInt-method}
\title{Seq method for the S4 class \linkS4class{RepoTimeInt}}
\usage{
Seq(x, y, Rot = TRUE, RotPer = "12")

\S4method{Seq}{RepoTimeInt}(x, y, Rot = TRUE, RotPer = "12")
}
\arguments{
\item{x}{Object of class \linkS4class{RepoTimeInt} with the initial time 
interval.}

\item{y}{Object of class \linkS4class{RepoTimeInt} with the final time 
interval.}

\item{Rot}{Logical vector of length 1 indicating whether periods with rotated
samples are to be included in the sequence (default) or not.}

\item{RotPer}{Character vector of length 1 with the numeric code of the time
period containing both unrotated and rotated samples.}
}
\value{
A list of objects of class \linkS4class{RepoTimeInt} ranging from
the initial time interval until the final time interval.
}
\description{
\code{Seq} generates a list of objects of class 
\linkS4class{RepoTimeInt} ranging from the initial and final time interval 
arguments.
}
\examples{
RepoPeriod1 <- newRepoTime('MM012015')
RepoPeriod2 <- newRepoTime('MM022016')
Seq(RepoPeriod1, RepoPeriod2)
 
}

