% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.RepoTimeInt.R
\name{show,RepoTimeInt-method}
\alias{show,RepoTimeInt-method}
\title{Show an object of class \linkS4class{RepoTimeInt}}
\usage{
\S4method{show}{RepoTimeInt}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{RepoTimeInt}.}
}
\value{
Object of class \code{NULL}.
}
\description{
\code{show} displays only the slot \code{Repo} of its input argument 
of class \linkS4class{RepoTimeInt}.
}
\details{
It is indeed the method \code{\link[methods]{show}} adapted to class 
\linkS4class{RepoTimeInt}.
}
\examples{
show(newRepoTime('TT12015'))

}
