% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Union.RepoTimeInt.R
\docType{methods}
\name{Union}
\alias{Union}
\alias{Union,RepoTimeInt-method}
\title{Union of two objects of class \linkS4class{RepoTimeInt}}
\usage{
Union(x, y)

\S4method{Union}{RepoTimeInt}(x, y)
}
\arguments{
\item{x}{Object of class \linkS4class{RepoTimeInt}.}

\item{y}{Object of class \linkS4class{RepoTimeInt}.}
}
\value{
Object of class \linkS4class{RepoTimeInt}.
}
\description{
\code{Union} takes two objects of class
\linkS4class{RepoTimeInt} and produces a new object of this class with the
slot \code{lubriInt} resulting from applying the set operation
\link[base]{union} to their corresponding slots \code{lubriInt}.
}
\details{
If argument \code{y} is missing, \code{union} is applied iteratively
on every successive component of the list of
\code{\link[lubridate]{Interval-class}} objects. If argument \code{y} is
not missing, the length of both arguments must be 1.

It is important to remind that \code{union} "fills in" the time gap between
their time interval arguments, if ever. It includes intervening time
intervals between the initial and final input time intervals.
}
\examples{
RepoPeriod1 <- newRepoTime('TT12015')
RepoPeriod2 <- newRepoTime('TT22015')
Union(RepoPeriod1, RepoPeriod2)

Months <- newRepoTime(paste0('MM', 10:12, '2014'))
Union(Months)
}

